initial work on using new InfiniteOpt backend API
